Create Collapse
Collapse is a game where the aim is to make a full line of a board using the several shapes provided. If a row is completely filled, the row will be removed and the remaining shapes above it will be "collapsed" one by one down through the empty space until they pile up on top of the shapes below. The stacking that occurs can produce a situation like at the beginning above, namely there are rows that are fully filled, which will then be removed again (after all collapse). This can happen repeatedly until there are no more rows that are completely filled.