Create Jawbreaker1
You've probably seen these computer games, maybe even played them. In front of you are a number of balls of various colors, of which the number of different colors is 5. The colors of the balls are numbered from 1 to 5. The balls are arranged in an M × N matrix and are placed randomly. Players can "pick up" a number of balls of the same color at once as long as the balls are in a series (explained in the next paragraph).