Create Quadtree1
Quadtree is a data structure that can be used to encode the contents of a binary image. Consider the following binary matrix, which is a matrix in which each element contains 2 possible values. The 0-valued parts are colored white and the 1-valued parts are grayed out.
If the matrix is divided into four square quadrants that have the same dimensions, namely the upper left, upper right, lower left, and lower right quadrants, then the upper left quadrant contains the same (homogeneous) values, while the other quadrants are not homogeneous. Homogeneous quadrants do not need to be divided again. For each quadrant that is not homogeneous, further division is carried out into 4 smaller quadrants, and so on, so that homogeneous submatrices will be formed.